refactor(api): update ThreadType imports and improve request handling
- Replaced imports of ThreadType from Enum.js to index.js across multiple API files for consistency.
- Enhanced request handling by removing unnecessary .toString() calls on service URLs in several API methods, improving code clarity and maintainability.
