+++ src/apis/addReaction.ts
@@ -1,5 +1,5 @@
 import { ZaloApiError } from "../Errors/ZaloApiError.js";
-import { ThreadType } from "../models/Enum.js";
+import { ThreadType } from "../models/index.js";
 import { Reactions } from "../models/Reaction.js";
 import { apiFactory } from "../utils.js";
