+++ src/apis/createReminder.ts
@@ -114,7 +114,7 @@ export const createReminderFactory = apiFactory<CreateReminderResponse>()((api, 
         const encryptedParams = utils.encodeAES(JSON.stringify(params));
         if (!encryptedParams) throw new ZaloApiError("Failed to encrypt params");
 
-        const response = await utils.request(serviceURL[type].toString(), {
+        const response = await utils.request(serviceURL[type], {
             method: "POST",
             body: new URLSearchParams({
                 params: encryptedParams,
