+++ src/apis/sendSticker.ts
@@ -57,7 +57,7 @@ export const sendStickerFactory = apiFactory<SendStickerResponse>()((api, ctx, u
         const encryptedParams = utils.encodeAES(JSON.stringify(params));
         if (!encryptedParams) throw new ZaloApiError("Failed to encrypt message");
 
-        const response = await utils.request(serviceURL[type].toString(), {
+        const response = await utils.request(serviceURL[type], {
             method: "POST",
             body: new URLSearchParams({
                 params: encryptedParams,
