+++ src/apis/sendVoice.ts
@@ -29,11 +29,7 @@ export const sendVoiceFactory = apiFactory<SendVoiceResponse>()((api, ctx, utils
      *
      * @throws ZaloApiError
      */
-    return async function sendVoice(
-        options: SendVoiceOptions,
-        threadId: string,
-        type: ThreadType = ThreadType.User,
-    ) {
+    return async function sendVoice(options: SendVoiceOptions, threadId: string, type: ThreadType = ThreadType.User) {
         let fileSize = null;
         let clientId = Date.now().toString();
 
@@ -46,26 +42,37 @@ export const sendVoiceFactory = apiFactory<SendVoiceResponse>()((api, ctx, utils
             throw new ZaloApiError(`Unable to get voice content: ${error?.message || error}`);
         }
 
-        const params: any = {
-            ttl: options.ttl ?? 0,
-            zsource: -1,
-            msgType: 3,
-            clientId: clientId,
-            msgInfo: JSON.stringify({
-                voiceUrl: options.voiceUrl,
-                m4aUrl: options.voiceUrl,
-                fileSize: fileSize ?? 0,
-            }),
-        };
+        const params =
+            type === ThreadType.User
+                ? {
+                      toId: threadId,
+                      ttl: options.ttl ?? 0,
+                      zsource: -1,
+                      msgType: 3,
+                      clientId: clientId,
+                      msgInfo: JSON.stringify({
+                          voiceUrl: options.voiceUrl,
+                          m4aUrl: options.voiceUrl,
+                          fileSize: fileSize ?? 0,
+                      }),
+                      imei: ctx.imei,
+                  }
+                : {
+                      grid: threadId,
+                      visibility: 0,
+                      ttl: options.ttl ?? 0,
+                      zsource: -1,
+                      msgType: 3,
+                      clientId: clientId,
+                      msgInfo: JSON.stringify({
+                          voiceUrl: options.voiceUrl,
+                          m4aUrl: options.voiceUrl,
+                          fileSize: fileSize ?? 0,
+                      }),
+                      imei: ctx.imei,
+                  };
 
-        if (type === 0) {
-            params.toId = threadId;
-            params.imei = ctx.imei;
-        } else if (type === 1) {
-            params.visibility = 0;
-            params.grid = threadId;
-            params.imei = ctx.imei;
-        } else {
+        if (type !== ThreadType.User && type !== ThreadType.Group) {
             throw new ZaloApiError("Thread type is invalid");
         }
